Revert "test_hybrid folder"
This reverts commit 786fe8576932038a6676b4ebbba3beca37674a04.
