--- test/test_hybrid/platformio.ini
@@ -1,36 +1,0 @@
-; PlatformIO Project Configuration File
-;
-;   Build options: build flags, source filter
-;   Upload options: custom upload port, speed and extra flags
-;   Library options: dependencies, extra library storages
-;   Advanced options: extra scripting
-;
-; Please visit documentation for the other options and examples
-; https://docs.platformio.org/page/projectconf.html
-
-[env]
-platform = espressif32
-board = esp32dev
-framework = arduino
-test_framework = googletest
-
-lib_deps =
-  adafruit/Adafruit MCP23017 Arduino Library@^2.3.2
-
-
-build_flags =
-  -std=gnu++17
-  -D CORE_DEBUG_LEVEL=5
-
-build_unflags =
-  -std=gnu++11
-
-[env:test]
-; Test related options
-lib_deps=
- ${env.lib_deps}
- ../..
-
-[env:ci]
-lib_deps=
- ${env.lib_deps}
